Chore: 알람 항목 컴포넌트 분리, 타입 분리
