+++ src/features/alarm/types/alarm.ts
@@ -1,0 +1,10 @@
+export type Alarm = {
+  id: number;
+  senderId: number;
+  receiverId: number;
+  postId: number;
+  senderNickname: string;
+  postTitle: string;
+  createdAt: string;
+  alarmType: string;
+};
